% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{ancova}
\alias{ancova}
\title{Analysis of Covariance}
\usage{
ancova(data, genotypes, replication)
}
\arguments{
\item{data}{traits to be analyzed}

\item{genotypes}{vector containing genotypes}

\item{replication}{vector containing replications}
}
\value{
ANCOVA, genotypic and phenotypic covariance
}
\description{
Analysis of Covariance
}
\examples{
data(vardata)
ancova(vardata[3:11],vardata$Genotypes,vardata$Replication)
}
