% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genvar.R
\name{gen.var}
\alias{gen.var}
\title{Estimation of Genetic Parameters}
\usage{
gen.var(data, genotypevector, replicationvector)
}
\arguments{
\item{data}{traits to be analyzed}

\item{genotypevector}{vector containing genotypes}

\item{replicationvector}{vector containig replications}
}
\value{
ANOVA, genotypic and phenotypic coefficient of variance, heritability, genetic advance and genetic advance as percentage of mean.
}
\description{
Estimation of Genetic Parameters
}
\examples{
data(vardata)
gen.var(vardata[3:11],vardata$Genotypes,vardata$Replication)
}
