% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencorr.R
\name{geno.corr}
\alias{geno.corr}
\title{Genotypic Correlation Analysis}
\usage{
geno.corr(data, genotypes, replication)
}
\arguments{
\item{data}{traits to be analyzed}

\item{genotypes}{vector containing genotypes}

\item{replication}{vector containing replications}
}
\value{
Genotypic correlation matrix
}
\description{
Genotypic Correlation Analysis
}
\examples{
data(vardata)
geno.corr(vardata[3:11],vardata$Genotypes,vardata$Replication)
}
