% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genopath.R
\name{geno.path}
\alias{geno.path}
\title{Genotypic Path Analysis}
\usage{
geno.path(dependent.var, independent.var, genotypes, replication)
}
\arguments{
\item{dependent.var}{trait to be used a dependent variable}

\item{independent.var}{traits to be used as an indpendent variables}

\item{genotypes}{vector containing genotpes}

\item{replication}{vector containing replications}
}
\value{
Direct effects, indirect effects and residual
}
\description{
Genotypic Path Analysis
}
\examples{
data(vardata)
# Grain yield is considered as a dependent variable
geno.path(vardata[11],vardata[3:10],vardata$Genotypes,vardata$Replication)
}
