% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenopath.R
\name{pheno.path}
\alias{pheno.path}
\title{Phenotypic Path Analysis}
\usage{
pheno.path(dependent.var, independent.var, genotypes, replication)
}
\arguments{
\item{dependent.var}{trait to be considered as a dependent variable}

\item{independent.var}{traits to be connsidered as an independent variables}

\item{genotypes}{vector containing genotypes}

\item{replication}{vector containing replicatons}
}
\value{
Direct effects, indirect effects and residual
}
\description{
Phenotypic Path Analysis
}
\examples{
data(vardata)
pheno.path(vardata[11],vardata[3:10],vardata$Genotypes,vardata$Replication)
}
