% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardata.R
\docType{data}
\name{vardata}
\alias{vardata}
\title{Variability Data}
\format{
The data has 11 columns and 120 rows
\describe{
\item{Genotypes}{40 genotypes}
\item{Replication}{3 replications}
\item{DFF}{Days to 50 per cent flowering}
\item{PH}{Plant height}
\item{PL}{Panicle length}
\item{PW}{Panicle weight}
\item{HI}{Harvest index}
\item{TW}{Test weight}
\item{MILL}{Milling percentage}
\item{HRR}{Head rice recovery}
\item{GY}{Grain Yield}
}
}
\usage{
vardata
}
\description{
The data consists of genotypes, replications and nine traits
}
\keyword{datasets}
